refactor(agentic-ai): move AgentContextSchemaMigration into a versioning package
Colocates the migration dispatch with VersionedAgentContextDeserializer,
its main caller, since both concern reading/binding a versioned
AgentContext; memory.conversation no longer fits after the earlier
rename, and top-level model was a step wide of the actual call site.

Also splits out ToolCallResultUpcaster as a nested class for the
tool-call-result content-lifting cluster (upcastMessages/
upcastToolCallResults/liftLegacyContent), separating it from the
schema-version dispatch that now lives in the outer class. Both
liftLegacyContent and upcastToolCallResults are now private -- neither
had a real external caller (the AWS AgentCore usage claimed by the old
javadoc doesn't exist in code), so their remaining unit coverage now
drives through upcastMessages, the one method with a real external
caller (CamundaDocumentConversationSerializer).
