+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/memory/conversation/document/CamundaDocumentConversationSerializer.java
@@ -10,9 +10,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
-import io.camunda.connector.agenticai.aiagent.memory.conversation.AgentContextSchemaMigration;
 import io.camunda.connector.agenticai.aiagent.memory.conversation.document.CamundaDocumentConversationContext.DocumentContent;
 import io.camunda.connector.agenticai.aiagent.model.AgentContext;
+import io.camunda.connector.agenticai.aiagent.model.versioning.AgentContextSchemaMigration;
 import io.camunda.connector.api.document.Document;
 import java.io.IOException;
 
@@ -46,7 +46,8 @@ public class CamundaDocumentConversationSerializer {
               .formatted(schemaVersion, AgentContext.CURRENT_SCHEMA_VERSION));
     }
     if (schemaVersion < AgentContext.CURRENT_SCHEMA_VERSION) {
-      AgentContextSchemaMigration.upcastMessages(tree.get("messages"), objectMapper);
+      AgentContextSchemaMigration.ToolCallResultUpcaster.upcastMessages(
+          tree.get("messages"), objectMapper);
     }
     return objectMapper.treeToValue(tree, DocumentContent.class);
   }
